refactor(player): unify package names to singular nouns
