+++ feature/mobile/player/src/main/java/com/flixclusive/feature/mobile/player/component/PlayerControls.kt
@@ -79,16 +79,16 @@ import com.flixclusive.feature.mobile.player.component.bottom.PlaybackSpeedSheet
 import com.flixclusive.feature.mobile.player.component.bottom.ResizeModeSheet
 import com.flixclusive.feature.mobile.player.component.center.CenterControls
 import com.flixclusive.feature.mobile.player.component.effect.NextEpisodeCountdownEffect
-import com.flixclusive.feature.mobile.player.component.episodes.EpisodesScreen
-import com.flixclusive.feature.mobile.player.component.gestures.BrightnessManager.Companion.rememberBrightnessManager
-import com.flixclusive.feature.mobile.player.component.gestures.PlayerGestureHandler
-import com.flixclusive.feature.mobile.player.component.gestures.PlayerGestureState.Companion.rememberPlayerGestureState
-import com.flixclusive.feature.mobile.player.component.gestures.SpeedBoostIndicator
-import com.flixclusive.feature.mobile.player.component.servers.ServersScreen
+import com.flixclusive.feature.mobile.player.component.episode.EpisodesScreen
+import com.flixclusive.feature.mobile.player.component.gesture.BrightnessManager.Companion.rememberBrightnessManager
+import com.flixclusive.feature.mobile.player.component.gesture.PlayerGestureHandler
+import com.flixclusive.feature.mobile.player.component.gesture.PlayerGestureState.Companion.rememberPlayerGestureState
+import com.flixclusive.feature.mobile.player.component.gesture.SpeedBoostIndicator
+import com.flixclusive.feature.mobile.player.component.server.ServersScreen
 import com.flixclusive.feature.mobile.player.component.snackbar.PlayerErrorSnackbar
 import com.flixclusive.feature.mobile.player.component.snackbar.PlayerSnackbar
-import com.flixclusive.feature.mobile.player.component.subtitles.SubtitleAndAudioScreen
-import com.flixclusive.feature.mobile.player.component.subtitles.SubtitleSyncScreen
+import com.flixclusive.feature.mobile.player.component.subtitle.SubtitleAndAudioScreen
+import com.flixclusive.feature.mobile.player.component.subtitle.SubtitleSyncScreen
 import com.flixclusive.feature.mobile.player.component.top.PlayerTopBar
 import com.flixclusive.feature.mobile.player.util.UiMode
 import com.flixclusive.model.film.FilmMetadata
