+++ feature/mobile/player/src/main/java/com/flixclusive/feature/mobile/player/component/episode/component/SeasonsRow.kt
@@ -1,4 +1,4 @@
-package com.flixclusive.feature.mobile.player.component.episodes.component
+package com.flixclusive.feature.mobile.player.component.episode.component
 
 import androidx.compose.foundation.gestures.Orientation
 import androidx.compose.foundation.layout.Arrangement
